app/requirements.txt
fastapi
uvicorn[standard]
requests>=2.28.0
pyyaml>=6.0
anthropic
